Bustle2.0 (#3)
* refactor utils

* refactor icon and toggleswitch

* refactor map, create toggle switch

* re-architect navigation to be root->drawer->tab->stack

* Full screen modal

* add with options restyle

* minor bug fixes

* map code refactor, add post support in redux, feed navigator, map polygons

* Upgrade all packages, feed header

* Adding changes to ListView

* Adding

* style changes, custom add icons, fix fb login

* Adding

* refactor form components and add button

* Adding

* new post infra and refactor for user data model

* add reply functionality

* bottomsheet, location filtering

* component restyles, rsvp button, tokenizer

* new event compose form

* event form tweaks, app recoloring, refactor navigators to simple stack

* validate event form

* Adding

* Adding finishing touches to lazy load

* restyle card, add in calendar saving

* overlay cards

* Adding profile additions

* Finished adding stuff to interested view

* Adding changes to intersted

* Calendar tab requeries from redux user, event bug fixes

* post feed style changes + search screen

* bug fixes w/ minor forum changes

* Adding forum changes(tag colors)

* fix tag ui and refactor sheet

* Anon on posts

* forum style fixes, add anonymous post, delete broken

* Adding some small changes

* Adding changes to ios and a lot of other features such as forums and profile

* simplify navigation, style fixes

* add edit button to event screen

* View other interested users, live sync event data

* style fixes, add in hosting to calendar view

* filter with event tags

* onboarding flow and cache events in redux

* ios and linkedin changes

* event detail button

* Adding ios changes

* saved event redux bug fix

* save event redux fix

* add cancel and remove interstitial

* Enabling remove and fixing ios

* interstitial

* interstitial

* ahdjskjadslg

* bugfix: new user flow and self user updates

* `Adding images and some bug fixes

* bugfix: loading prompt

* event filters fix and update event fix

* past events

* Adding some additional bug fixes

* Removing junk and reorganizing entire codebase

+++ models/Event.js
@@ -3,10 +3,10 @@ import { getLocation, validateLocation } from "../global/utils";
 import { firebase as f } from "@react-native-firebase/storage";
 import firestore, { firebase } from "@react-native-firebase/firestore";
 import { firebase as fire } from "@react-native-firebase/functions";
-import { GeoFirestore } from "geofirestore";
-import uuid from "uuid/v4";
+import { GeoFirestore, GeoTransaction } from "geofirestore";
 
 import { UserContext } from "../dataContainers/context";
+// import { getDefaultImage } from "../global/utils";
 
 // Event Class:
 // {
@@ -33,14 +33,27 @@ import { UserContext } from "../dataContainers/context";
 export default class EventModel {
   constructor() {}
 
-  static async get(filters, func) {
+  static async get(filters, onNext) {
     const store = firestore();
     const geofirestore = new GeoFirestore(store);
-    // const query = store.collection('events').orderBy('g').startAt(5)
     let query = geofirestore.collection("events");
+
+    if (filters.host) query = query.where("host.uid", "==", filters.host);
+    if (filters.categories?.length)
+      query = query.where("category", "in", filters.categories);
     if (filters.active) query = query.where("ended", "==", false);
-    if (filters.host) query = query.where("host", "==", filters.host);
-    if (func) query.onSnapshot(func);
+    if (filters.radius > 0)
+      query = query.near({
+        center: new firestore.GeoPoint(37.86835, -122.265),
+        radius: filters.radius,
+      });
+    if (filters.orderBy)
+      query = query.orderBy(
+        filters.orderBy,
+        filters.orderByDir ? filters.orderByDir : "asc"
+      );
+    if (filters.limit) query = query.limit(filters.limit);
+    if (onNext) query.onSnapshot(onNext, console.log);
   }
 
   static async getCollection() {
@@ -48,49 +61,40 @@ export default class EventModel {
   }
 
   static async remove(event) {
-    // const deletePhoto = fire.functions().httpsCallable("deletePhoto");
     const deleteEvent = fire.functions().httpsCallable("deleteEvent");
     await firestore()
       .collection("events")
       .doc(event.id)
       .delete()
-      .catch(function(error) {
+      .catch(function (error) {
         console.error("Error removing document: ", error);
       });
-    await firestore()
-      .collection("chats")
-      .doc(event.chatID)
-      .delete();
-    if (event.photoURL)
-      await f
-        .storage()
-        .refFromURL(event.photoURL)
-        .delete();
-    await deletePhoto({ eventID: event.id, photoURL: event.photoURL });
+    if (event.photoURL) await f.storage().refFromURL(event.photoURL).delete();
     for (let count; count < event.invited; count++) {
       const userID = event.invited[count];
       await deleteEvent({ uid: userID, eventID: event.id });
     }
   }
 
-  static async create(userID, data, events) {
+  static async create(user, data) {
     if (!data.name) throw new Error("Name not provided");
     const store = firestore();
     const geofirestore = new GeoFirestore(store);
 
-    const now = new Date();
+    const now = firebase.firestore.FieldValue.serverTimestamp();
     data.createdAt = now;
     data.date = data.date || now;
     data.time = data.time || now;
 
-    data.host = userID;
+    data.host = user;
     data.ended = false;
-    data.invited = [userID];
+    data.invited = {};
+    data.invited[user.uid] = true;
 
     const loc = await getLocation();
     if (data.location) {
       const { lat, lng } = data.location.geometry.location;
-      validateLocation(loc, lat, lng);
+      // validateLocation(loc, lat, lng);
       data.coordinates = new firestore.GeoPoint(lat, lng);
     } else {
       data.coordinates = new firestore.GeoPoint(
@@ -98,71 +102,68 @@ export default class EventModel {
         loc.coords.longitude
       );
     }
-    await store.runTransaction(async t => {
+
+    await geofirestore.runTransaction(async (t) => {
+      const transaction = new GeoTransaction(t);
+      const eventRef = store.collection("events").doc();
       // upload image
       if (data.image) {
-        const fileName = `${event.id}`;
-        const snapshot = await f
+        await f
           .storage()
-          .ref(`events/${data.category}/${fileName}`)
-          .put(data.image, {
-            customMetadata: { uid: userID, eventID: event.id }
-          })
-          .then();
-        data.photoURL = snapshot.ref.getDownloadURL();
+          .ref(`events/${data.category}/${eventRef}`)
+          .putFile(data.image);
+        data.photoURL = await f
+          .storage()
+          .ref(`events/${data.category}/${eventRef}`)
+          .getDownloadURL();
+        data.photoID = eventRef;
         delete data.image;
       }
-
       // upload event
-      const eventRef = await geofirestore.collection("events").add(data);
+      transaction.set(eventRef, data);
 
-      // denormed update on user
-      events.push(eventRef.id);
-      await store
-        .collection("users")
-        .doc(userID)
-        .update({ events: events });
+      // // denormed update on user
+      return transaction.native.update(
+        store
+          .collection("users")
+          .doc(user.uid)
+          .collection("private")
+          .doc("profile"),
+        {
+          events: firestore.FieldValue.arrayUnion(eventRef.id),
+        }
+      );
     });
   }
 
   static async update(eventID, data) {
-    // const deletePhoto = fire.functions().httpsCallable("deletePhoto");
-    // const uploadPhoto = fire.functions().httpsCallable("uploadPhoto");
     if (!data.name) throw new Error("Name not provided");
     if (data.location) {
       const loc = await getLocation();
       const { lat, lng } = data.location.geometry.location;
-      validateLocation();
+      validateLocation(loc, lat, lng);
     }
     const store = firestore();
     const geofirestore = new GeoFirestore(store);
-    if (data.image) {
-      if (data.photoURL)
+
+    await geofirestore.runTransaction(async (t) => {
+      const transaction = new GeoTransaction(t);
+      if (data.image) {
+        if (data.photoURL) {
+          await f.storage().refFromURL(data.photoURL).delete();
+        }
         await f
           .storage()
-          .refFromURL(event.photoURL)
-          .delete();
-      //   await deletePhoto({ eventID: event.id, photoURL: event.photoURL });
-      const fileName = `${uuid()}`;
-      await f
-        .storage()
-        .ref(`events/${data.category}/${fileName}`)
-        .put(data.image, {
-          customMetadata: { uid: data.host, eventID: event.id }
-        })
-        .on(f.storage.TaskEvent.STATE_CHANGED, async snapshot => {
-          if (snapshot.state === firebase.storage.TaskState.SUCCESS) {
-            const photoURL = await snapshot.ref.getDownloadURL();
-            // await uploadPhoto({ eventID: event.id, photoURL: photoURL });
-          }
-        });
-    }
-    data.photoURL = data.image;
-    delete data.image;
-    await geofirestore
-      .collection("events")
-      .doc(eventID)
-      .update(data);
+          .ref(`events/${data.category}/${eventID}`)
+          .putFile(data.image);
+        data.photoURL = await f
+          .storage()
+          .ref(`events/${data.category}/${eventID}`)
+          .getDownloadURL();
+        delete data.image;
+      }
+      transaction.update(store.collection(collection).doc(eventID), data);
+    });
   }
 }
